requirements.txt
PyMuPDF==1.22.3
pdf2image
pytesseract
opencv-python
numpy
Pillow
langdetect
